requirements.txt
watchdog
redis
rq
sqlalchemy
mysql-connector-python